created searchResultsPage & searchResults component -- trying to test whether props (searchTerm) is being sent to where it needs to go
