+++ my-express-app/client/vite-project/src/App.jsx
@@ -1,5 +1,6 @@
 import { useState } from "react";
 import Search from "./Components/Search";
+import SearchResultsPage from "./Pages/SearchResultsPage";
 
 import "./App.css"; // This file is empty currently.
 
@@ -8,7 +9,7 @@ function App() {
     <div className="App">
       <div>
         <h1>This is the front end.</h1>
-        <Search />
+        <SearchResultsPage />
       </div>
     </div>
   );
