+++ my-express-app/client/vite-project/src/Components/Search.jsx
@@ -1,36 +1,36 @@
 import React from "react";
 import { useState } from "react";
+import { useNavigate } from "react-router-dom";
 
-function Search() {
+function Search({ sendSearchTermCB }) {
   const [loading, setLoading] = useState(false);
   const [searchTerm, setSearchTerm] = useState("");
 
-  //Function for full search with API -- not working yet. getting 404 error
-  const searchPodcast = async (searchTerm) => {
-    setLoading(true);
-    let options = {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ searchTerm: searchTerm }),
-    };
-    try {
-      let results = await fetch(`/api/search`, options);
-      let data = await results.json();
-      console.log(data);
-      return data; // Returns array of 10 podcast episodes that meet search criteria
-    } catch (err) {
-      console.log(err);
-    }
-    setLoading(false);
-  };
+  //   //Function for full search with API -- not working yet. getting 404 error
+  //   const searchPodcast = async (searchTerm) => {
+  //     setLoading(true);
+  //     let options = {
+  //       method: "POST",
+  //       headers: {
+  //         "Content-Type": "application/json",
+  //       },
+  //       body: JSON.stringify({ searchTerm: searchTerm }),
+  //     };
+  //     try {
+  //       let results = await fetch(`/api/search`, options);
+  //       let data = await results.json();
+  //       console.log(data);
+  //       return data; // Returns array of 10 podcast episodes that meet search criteria
+  //     } catch (err) {
+  //       console.log(err);
+  //     }
+  //     setLoading(false);
+  //   };
 
   const handleSubmit = (e) => {
     e.preventDefault();
-    searchPodcast(searchTerm);
-    console.log(searchTerm); //Working
-    setSearchTerm("");
+    sendSearchTermCB(searchTerm);
+    // useNavigate("/results"); //This page needs to be set up in the router area.
   };
 
   return (
